+++ panov.ilya/P1/nchar.hpp
@@ -7,7 +7,7 @@ namespace panov
     public:
       Nchar();
       void operator()(int number);
-      int operator()() const;
+      size_t operator()() const;
     private:
       int count;
       int first;
